Add image resources
This allows the usage of PNG sprites with the new procedure
blit-image. blit-image can be used on resources defined with the new
define-resource procedure. Images can be scaled (around the original
center) and provided with an offset for moving tiles automatically
when placing. This is used for tilesets that have an "empty"
background above for stackable items.