Add mod keys to keyhandling and translate properly
This commit changes the event format for key events:

(key pressed-state modifier key)

Where key can be a character or a symbol describing the key. modifier is
a list that might be empty containing the modifiers for a key.

Note that press/release events of the modifier keys alone will also be
emitted.

Existing code needs to be changed!